pyproject.toml
[tool.poetry]
authors = ["Magnus Dalen Kvalevåg <magnus.kvalevag@gmail.com>"]
description = ""
name = "edesdetectrl"
version = "0.1.0"

[tool.poetry.dependencies]
SimpleITK = "^2.1.1"
absl-py = "^0.12.0"
chex = "^0.0.8"
dm-acme = "^0.3.0"
dm-haiku = "^0.0.5"
dm-reverb = "^0.6.1"
dm-sonnet = "^2.0.0"
gym = "^0.20.0"
jax = "<=0.2.25"
jaxlib = {url = "https://storage.googleapis.com/jax-releases/cuda111/jaxlib-0.1.71+cuda111-cp38-none-manylinux2010_x86_64.whl"}
libclang = "<12.0.0"
mlflow = "^1.23.0"
nevergrad = "^0.4.3.post9"
opencv-python = "^4.5.5.62"
optax = "^0.0.9"
pandas = "^1.3.5"
pip = "^21.3"
python = ">=3.8,<3.11"
rlax = "^0.0.4"
scikit-learn = "^0.24.2"
tensorflow = "^2.7.0"
tensorflow-probability = "^0.15.0"
toml = "^0.10.2"
torch = "^1.10.1"

[tool.poetry.dev-dependencies]
PySimpleGUI = "^4.56.0"
black = "^20.8b1"
ipykernel = "^6.7.0"
jurigged = "^0.4.1"
matplotlib = "^3.5.1"
poetryup = "^0.4.3"
pytest = "^5.4.3"
gif = "^3.0.0"
hypothesis = "^6.36.2"

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core>=1.0.0"]
